requirements.txt
bcrypt==4.0.1
cryptography
email-validator
fastapi==0.111.0
passlib[bcrypt]==1.7.4
psycopg2-binary
pydantic
firebase-admin==6.4.0
pydantic_core
PyJWT
SQLAlchemy
python-dotenv
uvicorn
python-multipart
beautifulsoup4==4.14.3
pika
fastapi_sqlalchemy==0.2.1
loguru
alembic==1.13.1
itsdangerous==2.2.0
sendgrid==6.11.0
requests
boto3

